+++ console/migrations/m150104_153617_create_article_table.php
@@ -1,4 +1,5 @@
 <?php
+
 use yii\db\Schema;
 use yii\db\Migration;
 
@@ -8,19 +9,18 @@ class m150104_153617_create_article_table extends Migration
     {
         $tableOptions = null;
 
-        if ($this->db->driverName === 'mysql') 
-        {
+        if ($this->db->driverName === 'mysql') {
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
         }
 
         $this->createTable('{{%article}}', [
-            'id' => Schema::TYPE_PK,
-            'user_id' => Schema::TYPE_INTEGER . ' NOT NULL',
-            'title' => Schema::TYPE_STRING . ' NOT NULL',
-            'summary' => Schema::TYPE_TEXT . ' NOT NULL',
-            'content' => Schema::TYPE_TEXT . ' NOT NULL',
-            'status' => Schema::TYPE_INTEGER . ' NOT NULL',
-            'category' => Schema::TYPE_INTEGER . ' NOT NULL',          
+            'id'         => Schema::TYPE_PK,
+            'user_id'    => Schema::TYPE_INTEGER . ' NOT NULL',
+            'title'      => Schema::TYPE_STRING . ' NOT NULL',
+            'summary'    => Schema::TYPE_TEXT . ' NOT NULL',
+            'content'    => Schema::TYPE_TEXT . ' NOT NULL',
+            'status'     => Schema::TYPE_INTEGER . ' NOT NULL',
+            'category'   => Schema::TYPE_INTEGER . ' NOT NULL',
             'created_at' => Schema::TYPE_INTEGER . ' NOT NULL',
             'updated_at' => Schema::TYPE_INTEGER . ' NOT NULL',
             'FOREIGN KEY (user_id) REFERENCES {{%user}}(id)
@@ -31,6 +31,12 @@ class m150104_153617_create_article_table extends Migration
 
     public function down()
     {
-        $this->dropTable('{{%article}}');
+        if ( !YII_ENV_PROD ) {
+            $this->dropTable('{{%article}}');
+        } else {
+            echo "WARNING: " . __CLASS__ . " cannot be reverted.\n";
+
+            return false;
+        }
     }
 }
